1. crypto logging level can be set using property logging.crypto 2. grpc connections opened by the crypto layer are logged in more detail
