+++ openchain/crypto/crypto_test.yaml
@@ -96,7 +96,7 @@ logging:
     #   warning:main,db=debug:chaincode=info       - Override default WARNING in main,db,chaincode
     #   chaincode=info:main=debug:db=debug:warning - Same as above
     peer:      debug
-    crypto:    warning
+    crypto:    debug
     status:    warning
     stop:      warning
     login:     warning
